old-II/requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
jsonschema>=4.20.0

# Environment & Configuration
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Cache & Message Queue
redis>=5.0.0
celery>=5.3.0

# LLM Providers
google-generativeai>=0.3.0
litellm>=1.20.0

# Utilities
langdetect>=1.0.9

# Monitoring & Observability
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
python-json-logger>=2.0.7

# Distributed Tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0

# Load Testing
locust>=2.42.0

# System Monitoring
psutil>=5.9.0
